Applied patch #6668
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/libobjc/trunk@27082 72102866-910b-0410-8b05-ffd578937521

+++ thr-win32.c
@@ -72,12 +72,12 @@ __objc_thread_detach(void (*func)(void *arg), void *arg)
                                    arg, 0, &thread_id)))
 #else 
   // According to MSDN documentation threads which use libc functions should call _beginthreadex not CreateThread
-  if ((HANDLE)-1 == (win32_handle = (HANDLE)_beginthreadex(NULL, 0, (void*)func, arg, 0, (int*)&thread_id)))
+  if ((HANDLE)-1 == (win32_handle = (HANDLE)_beginthreadex(NULL, 0, (void*)func, arg, 0, (unsigned*)&thread_id)))
 #endif
     thread_id = 0;
     
   CloseHandle((HANDLE)win32_handle);
-  return (objc_thread_t)thread_id;
+  return (objc_thread_t)(size_t) thread_id;
 }
 
 /* Set the current thread's priority. */
@@ -162,7 +162,7 @@ __objc_thread_exit(void)
 objc_thread_t
 __objc_thread_id(void)
 {
-  return (objc_thread_t)GetCurrentThreadId();
+  return (objc_thread_t)(size_t) GetCurrentThreadId();
 }
 
 /* Sets the thread's local storage pointer. */
